pyproject.toml
[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[project]
name = "hello-server"
version = "0.1.0"
description = "An MCP server built with Smithery"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

requires-python = ">=3.10"
dependencies = [
    "mcp>=1.15.0",
    "smithery>=0.4.2",
    "supabase>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dev = "smithery.cli.dev:main" # Run the MCP server in development mode
start = "smithery.cli.start:main" # Run the MCP server in production mode
playground = "smithery.cli.playground:main" # Run server with interactive testing playground

[tool.smithery]
server = "hello_server.server:create_server"
log_level = "warning" # Production log level (optional, defaults to "warning")
